feature/internet-connection-checker Fix: Update NetworkDataSourceImpl
Update the NetworkDataSourceImpl to pass the lifecycle and simplify the manageNetworkManager
